Move into a mixin to reuse
